task 3: name conventions fixed
