+++ homework-assignments-1/task-3/book-details.interface.ts
@@ -1,0 +1,5 @@
+export interface IBookDetails {
+  title: string;
+  author: string;
+  isbn?: number;
+}
