--- homework-assignments-1/task-3/book.interface.ts
@@ -1,5 +1,0 @@
-export interface IBook {
-    title: string;
-    author: string;
-    isbn: number;
-}
