+++ homework-assignments-1/task-3/library.service.ts
@@ -1,18 +1,20 @@
-import { IBook } from "./book.interface";
+import { IBookDetails } from "./book-details.interface";
 import { ILibrary } from "./library.interface";
 
 export class Library implements ILibrary {
-    private books: IBook[];
-    constructor(externalBookStore?: IBook[]) {
-        this.books = externalBookStore || [];
-    }
-    addBook(book: IBook): boolean {
-        this.books.push(book);
-        return true;
-    }
+  private books: IBookDetails[];
+  constructor(externalBookStore?: IBookDetails[]) {
+    this.books = externalBookStore || [];
+  }
+  addBook(book: IBookDetails): boolean {
+    this.books.push(book);
+    return true;
+  }
 
-    removeBookByISBN(isbn: number): boolean {
-        this.books = this.books.filter(book => { book.isbn != isbn });
-        return true;
-    }
+  removeBookByISBN(isbn: number): boolean {
+    this.books = this.books.filter((book) => {
+      book.isbn != isbn;
+    });
+    return true;
+  }
 }
